perf(test): voeg de twee circuit-breaker-tests samen tot één instantie
Beide tests hadden een eigen leeg profiel, puur om een schone applicatie-instantie af
te dwingen — breaker-state is application-scoped en blijft na een trip nog 5 seconden
open. Twee lege profielen kostten twee starts plus twee database-containers voor
dezelfde isolatie-eis; het zijn nu drie testmethodes in één klasse met één profiel.

De volgorde is functioneel en staat vast met @Order. De opslagtest gaat over HTTP en
valideert dus vóór het opslaan: draait die ná de Profiel-storingstest, dan fast-failt
hij op het inmiddels open validatiecircuit en wordt hij groen zonder het opslagcircuit
ooit te raken. Daarom opslag eerst, dan de weigering (die niets opent), en als laatste
de storing.

Gemeten over de hele magazijn-module, lokaal, schone build:

| | vóór | na |
|---|---|---|
| applicatie-starts | 16 | 15 |
| boottijd | 75,3s | 68,4s |
| database-containers | 15 | 14 |
| moduletijd | 194s | 186s |

379 tests, ongewijzigd, allemaal groen.
